Basic website layout
Here I modify the index.html file to give a bare-bones layout for the
physics and math tutorial topics. U-substitution integrals will be the
first topic written, and a link to the (unfinished) u-substitution html
is given (need to verify it works when pushed).

Also created a global.css file which will be expanded later.